// src/entity/Profile/model/types/profile.ts
import { Currency } from 'entity/Currency/model/types/currency';
import { Country } from 'entity/Country';

export enum ValidationProfileError {
  INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
  INCORRECT_AGE = 'INCORRECT_AGE',
  INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
  NO_DATA = 'NO_DATA',
  SERVER_ERROR = 'SERVER_ERROR',
}

export interface Profile {
  firstname?: string,
  lastname?: string,
  age?: number,
  currency?: Currency,
  country?: Country,
  city?: string,
  username?: string,
  avatar?: string
}

export interface ProfileSchema {
  data?: Profile;
  form?: Profile; // edited data
  isLoading: boolean;
  error?: string;
  readonly: boolean;
  validateErrors?: ValidationProfileError[];
}